fix(product): Price VO의 정밀도 손실 문제 해결

+++ src/main/java/com/livebidding/server/product/domain/vo/Price.java
@@ -42,7 +42,7 @@ public class Price {
         }
     }
 
-    public Long getValue() {
-        return value.longValue();
+    public BigDecimal getValue() {
+        return value;
     }
 }
